Implement unit tests for OrderViewModel
- refactor some things to adequate for tests

+++ app/src/main/java/com/carlosolimpio/minhasvendas/data/local/entity/OrderEntity.kt
@@ -3,13 +3,13 @@ package com.carlosolimpio.minhasvendas.data.local.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import androidx.room.TypeConverter
-import kotlinx.serialization.json.Json
 import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
 
 @Entity(tableName = "order_table")
 data class OrderEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val clientName: String,
     val items: List<ItemEntity>
 )
